[LEMS-2986/answerless-group] add stories
