+++ packages/perseus/src/widgets/group/group.stories.tsx
@@ -2,7 +2,11 @@ import {generateTestPerseusItem} from "@khanacademy/perseus-core";
 
 import {ServerItemRendererWithDebugUI} from "../../../../../testing/server-item-renderer-with-debug-ui";
 
-import {question1} from "./group.testdata";
+import {
+    getFullGroupTestItem,
+    getSplitGroupTestItem,
+    question1,
+} from "./group.testdata";
 
 import type {Meta, StoryObj} from "@storybook/react";
 
@@ -17,5 +21,17 @@ type Story = StoryObj<typeof ServerItemRendererWithDebugUI>;
 export const Question1: Story = {
     args: {
         item: generateTestPerseusItem({question: question1}),
+    },
+};
+
+export const Answerful: Story = {
+    args: {
+        item: getFullGroupTestItem(),
+    },
+};
+
+export const Answerless: Story = {
+    args: {
+        item: getSplitGroupTestItem(),
     },
 };
